feat: init DID registry
