+++ src/iam.ts
@@ -119,111 +119,7 @@ export class IAM extends IAMBase {
 
     // DID DOCUMENT
 
-    /**
-     * getDidDocument
-     *
-     * @returns whole did document if connected, if not returns null
-     *
-     */
-    async getDidDocument({
-        did = this._did,
-        includeClaims = true,
-    }: { did?: string; includeClaims?: boolean } | undefined = {}) {
-        if (this._cacheClient && did) {
-            try {
-                const didDoc = await this._cacheClient.getDidDocument({ did, includeClaims });
-                return {
-                    ...didDoc,
-                    service: didDoc.service as (IServiceEndpoint & ClaimData)[],
-                };
-            } catch (err) {
-                if ((err as AxiosError).response?.status === 401) {
-                    throw err;
-                }
-                console.log(err);
-            }
-        }
-
-        if (did && this._resolver) {
-            const document = await this._resolver.read(did);
-            return {
-                ...document,
-                service: includeClaims
-                    ? await this.downloadClaims({
-                          services: document.service && document.service.length > 0 ? document.service : [],
-                      })
-                    : [],
-            };
-        }
-        throw new Error(ERROR_MESSAGES.USER_NOT_LOGGED_IN);
-    }
-
-    /**
-     * @param options Options to connect with blockchain
-     *
-     * @param options.didAttribute Type of document to be updated
-     *
-     * @param options.data New attribute value
-     * @param options.did Asset did to be updated
-     * @param options.validity Time (s) for the attribute to expire
-     *
-     * @description updates did document based on data provided
-     * @returns true if document is updated successfuly
-     *
-     */
-    async updateDidDocument(options: {
-        didAttribute: DIDAttribute;
-        data: IUpdateData;
-        did?: string;
-        validity?: number;
-    }): Promise<boolean> {
-        const { didAttribute, data, validity, did } = options;
-
-        if (!did) {
-            if (!this._document) {
-                throw new Error(ERROR_MESSAGES.DID_DOCUMENT_NOT_INITIALIZED);
-            }
-            const updated = await this._document.update(didAttribute, data, validity);
-            return Boolean(updated);
-        }
-
-        if (!this._didSigner) {
-            throw new Error(ERROR_MESSAGES.SIGNER_NOT_INITIALIZED);
-        }
-
-        const updateData: IUpdateData = {
-            algo: Algorithms.Secp256k1,
-            encoding: Encoding.HEX,
-            ...data,
-        };
-
-        const operator = new ProxyOperator(this._didSigner, this._registrySetting, addressOf(did));
-        const update = await operator.update(did, didAttribute, updateData);
-
-        return Boolean(update);
-    }
-
-    /**
-     * revokeDidDocument
-     *
-     * @description revokes did document
-     * @returns information (true/false) if the DID document was revoked
-     *
-     */
-    async revokeDidDocument(): Promise<boolean> {
-        if (this._document) {
-            await this._document.deactivate();
-            return true;
-        }
-        throw new Error(ERROR_MESSAGES.DID_DOCUMENT_NOT_INITIALIZED);
-    }
-
-    async decodeJWTToken({ token }: { token: string }) {
-        if (!this._jwt) {
-            throw new Error(ERROR_MESSAGES.JWT_NOT_INITIALIZED);
-        }
-        return this._jwt.decode(token);
-    }
+    
 
     /// ROLES
